nexus-light-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
sqlalchemy
requests
pytest-mock
